[AllBundle] Cleanup sf 3.4 compatibility layers
