--- .build/stubs/RoleInterface.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace Symfony\Component\Security\Core\Role;
-
-//NEXT_MAJOR: when sf 3.4 support is dropped this stub can also be removed
-interface RoleInterface
-{
-
-    /**
-     * @return string|null
-     */
-    public function getRole();
-
-}
