+++ composer.json
@@ -33,6 +33,7 @@
         "symfony/http-foundation": "^4.4",
         "symfony/http-kernel": "^4.4",
         "symfony/inflector": "^4.4",
+        "symfony/mailer": "^4.4",
         "symfony/monolog-bridge": "^4.4",
         "symfony/options-resolver": "^4.4",
         "symfony/property-access": "^4.4",
