+++ src/Kunstmaan/AdminBundle/EventListener/AdminLocaleListener.php
@@ -9,7 +9,6 @@ use Symfony\Component\HttpKernel\Event\RequestEvent;
 use Symfony\Component\HttpKernel\KernelEvents;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
-use Symfony\Component\Translation\TranslatorInterface as LegacyTranslatorInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
@@ -23,7 +22,7 @@ class AdminLocaleListener implements EventSubscriberInterface
     private $tokenStorage;
 
     /**
-     * @var TranslatorInterface|LegacyTranslatorInterface
+     * @var TranslatorInterface
      */
     private $translator;
 
@@ -43,17 +42,11 @@ class AdminLocaleListener implements EventSubscriberInterface
     private $adminRouteHelper;
 
     /**
-     * @param TranslatorInterface|LegacyTranslatorInterface $translator
-     * @param string                                        $defaultAdminLocale
-     * @param string                                        $providerKey        Firewall name to check against
+     * @param string $defaultAdminLocale
+     * @param string $providerKey        Firewall name to check against
      */
-    public function __construct(TokenStorageInterface $tokenStorage, /* TranslatorInterface|LegacyTranslatorInterface */ $translator, AdminRouteHelper $adminRouteHelper, $defaultAdminLocale, $providerKey = 'main')
+    public function __construct(TokenStorageInterface $tokenStorage, TranslatorInterface $translator, AdminRouteHelper $adminRouteHelper, $defaultAdminLocale, $providerKey = 'main')
     {
-        // NEXT_MAJOR Add "Symfony\Contracts\Translation\TranslatorInterface" typehint when sf <4.4 support is removed.
-        if (!$translator instanceof \Symfony\Contracts\Translation\TranslatorInterface && !$translator instanceof LegacyTranslatorInterface) {
-            throw new \InvalidArgumentException(sprintf('The "$translator" parameter should be instance of "%s" or "%s"', TranslatorInterface::class, LegacyTranslatorInterface::class));
-        }
-
         $this->translator = $translator;
         $this->tokenStorage = $tokenStorage;
         $this->defaultAdminLocale = $defaultAdminLocale;
