+++ src/Kunstmaan/AdminBundle/Helper/Security/Acl/Permission/PermissionAdmin.php
@@ -20,7 +20,6 @@ use Symfony\Component\Security\Acl\Model\MutableAclInterface;
 use Symfony\Component\Security\Acl\Model\MutableAclProviderInterface;
 use Symfony\Component\Security\Acl\Model\ObjectIdentityRetrievalStrategyInterface;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
-use Symfony\Component\Security\Core\Role\RoleInterface;
 
 /**
  * Helper to manage the permissions on a certain entity
@@ -150,13 +149,13 @@ class PermissionAdmin
     /**
      * Get permission for specified role.
      *
-     * @param RoleInterface|string $role
+     * @param \Symfony\Component\Security\Core\Role\Role|string $role
      *
      * @return MaskBuilder|null
      */
     public function getPermission($role)
     {
-        if ($role instanceof RoleInterface || $role instanceof \Symfony\Component\Security\Core\Role\Role) {
+        if ($role instanceof \Symfony\Component\Security\Core\Role\Role) {
             $role = $role->getRole();
         }
         if (isset($this->permissions[$role])) {
