+++ src/Kunstmaan/AdminBundle/KunstmaanAdminBundle.php
@@ -6,7 +6,6 @@ use Kunstmaan\AdminBundle\DependencyInjection\Compiler\AddLogProcessorsCompilerP
 use Kunstmaan\AdminBundle\DependencyInjection\Compiler\AdminPanelCompilerPass;
 use Kunstmaan\AdminBundle\DependencyInjection\Compiler\DataCollectorPass;
 use Kunstmaan\AdminBundle\DependencyInjection\Compiler\EnablePermissionsPass;
-use Kunstmaan\AdminBundle\DependencyInjection\Compiler\InjectUntrackedTokenStorageCompilerPass;
 use Kunstmaan\AdminBundle\DependencyInjection\Compiler\MenuCompilerPass;
 use Kunstmaan\AdminBundle\DependencyInjection\KunstmaanAdminExtension;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -28,7 +27,6 @@ class KunstmaanAdminBundle extends Bundle
         $container->addCompilerPass(new AdminPanelCompilerPass());
         $container->addCompilerPass(new AddLogProcessorsCompilerPass());
         $container->addCompilerPass(new DataCollectorPass());
-        $container->addCompilerPass(new InjectUntrackedTokenStorageCompilerPass());
         $container->addCompilerPass(new EnablePermissionsPass());
 
         $container->registerExtension(new KunstmaanAdminExtension());
